requirements.txt
dnspython

rich

requests

colorama

collections

beautifulsoup4

cryptography

tenacity

datetime

pathlib

socks